+++ variables.tf
@@ -29,6 +29,12 @@ variable "names" {
   type        = list(string)
 }
 
+variable "randomize_suffix" {
+  description = "Adds an identical, but randomized 4-character suffix to all bucket names"
+  type        = bool
+  default     = false
+}
+
 variable "location" {
   description = "Bucket location."
   type        = string
@@ -38,7 +44,7 @@ variable "location" {
 variable "storage_class" {
   description = "Bucket storage class."
   type        = string
-  default     = "MULTI_REGIONAL"
+  default     = "STANDARD"
 }
 
 variable "force_destroy" {
@@ -182,7 +188,10 @@ variable "lifecycle_rules" {
     # - with_state - (Optional) Match to live and/or archived objects. Supported values include: "LIVE", "ARCHIVED", "ANY".
     # - matches_storage_class - (Optional) Comma delimited string for storage class of objects to satisfy this condition. Supported values include: MULTI_REGIONAL, REGIONAL, NEARLINE, COLDLINE, STANDARD, DURABLE_REDUCED_AVAILABILITY.
     # - num_newer_versions - (Optional) Relevant only for versioned objects. The number of newer versions of an object to satisfy this condition.
+    # - custom_time_before - (Optional) A date in the RFC 3339 format YYYY-MM-DD. This condition is satisfied when the customTime metadata for the object is set to an earlier date than the date used in this lifecycle condition.
     # - days_since_custom_time - (Optional) The number of days from the Custom-Time metadata attribute after which this condition becomes true.
+    # - days_since_noncurrent_time - (Optional) Relevant only for versioned objects. Number of days elapsed since the noncurrent timestamp of an object.
+    # - noncurrent_time_before - (Optional) Relevant only for versioned objects. The date in RFC 3339 (e.g. 2017-06-13) when the object became nonconcurrent.
     condition = map(string)
   }))
   description = "List of lifecycle rules to configure. Format is the same as described in provider documentation https://www.terraform.io/docs/providers/google/r/storage_bucket.html#lifecycle_rule except condition.matches_storage_class should be a comma delimited string."
@@ -190,15 +199,21 @@ variable "lifecycle_rules" {
 }
 
 variable "cors" {
-  description = "Map of maps of mixed type attributes for CORS values. See appropriate attribute types here: https://www.terraform.io/docs/providers/google/r/storage_bucket.html#cors"
-  type        = any
-  default     = {}
+  description = "Set of maps of mixed type attributes for CORS values. See appropriate attribute types here: https://www.terraform.io/docs/providers/google/r/storage_bucket.html#cors"
+  type        = set(any)
+  default     = []
 }
 
 variable "website" {
-  type        = any
+  type        = map(any)
   default     = {}
   description = "Map of website values. Supported attributes: main_page_suffix, not_found_page"
+}
+
+variable "retention_policy" {
+  type        = any
+  default     = {}
+  description = "Map of retention policy values. Format is the same as described in provider documentation https://www.terraform.io/docs/providers/google/r/storage_bucket#retention_policy"
 }
 
 variable "logging" {