+++ versions.tf
@@ -25,7 +25,7 @@ terraform {
   }
 
   provider_meta "google" {
-    module_name = "blueprints/terraform/terraform-google-cloud-storage/v2.1.0"
+    module_name = "blueprints/terraform/terraform-google-cloud-storage/v3.0.0"
   }
 
 }